requirements.txt
attrs==22.1.0
black==22.8.0
cffi==1.15.1
click==8.1.3
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.3
tomli==2.0.1
